+++ src/main/java/com/lyhour/java/study/phone_shop/repository/ModelRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.lyhour.java.study.phone_shop.entity.Model;
 
-public interface ModelRepository extends JpaRepository<Model, Integer> {
-	List<Model> findModelByBrandId(Integer brandId);
+public interface ModelRepository extends JpaRepository<Model, Long> {
+	List<Model> findModelByBrandId(Long brandId);
 	
 
 }
